# tech/public/openapi.yaml
openapi: 3.0.3
info:
  title: Surftorial API
  description: 서핑 온라인 강좌 플랫폼 Surftorial의 REST API 명세
  version: 1.0.0
  contact:
    name: Surftorial Team
    url: https://surftorial.com
servers:
  - url: https://surftorial.com/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase Auth 세션 쿠키
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase JWT (Phase 2)
...
paths:
...
  /api/courses/{id}/reviews:
    get:
      tags: [Reviews]
      summary: 강의 리뷰 목록 조회 (Public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 리뷰 목록 + 평균 평점
    post:
      tags: [Reviews]
      summary: 강의 리뷰 작성 (수강생만)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
                  maxLength: 1000
      responses:
        '201':
          description: 리뷰 작성 성공
        '403':
          description: 수강 중인 강의만 리뷰 가능

  # ===== PAYMENTS =====
...